chore: add sdkdata field to indicate that paypal v6 is supported
