+++ packages/lib/src/components/PayPal/stories/PaypalExpressWithSessions.stories.tsx
@@ -7,7 +7,7 @@ import { getDeliveryMethods, getSelectedDeliveryMethodAmount } from './paypal-st
 import { PayPal } from '../../..';
 
 const meta: Meta = {
-    title: 'Components/Wallets/Paypal',
+    title: 'Components/Wallets/Paypal/Express',
     tags: ['no-automated-visual-test']
 };
 export default meta;
@@ -24,7 +24,7 @@ const SHOPPER_LOCALE = 'en-US';
  * This must be implemented on the backend side by the merchant for safety reasons.
  */
 
-export const ExpressWithSessionsFlow: StoryObj = {
+export const WithSessionsFlow: StoryObj = {
     render: () => {
         return <Component />;
     },
